Something broken

+++ app/src/main/java/com/example/sensdataudp/MainActivity.java
@@ -75,7 +75,7 @@ public class MainActivity extends AppCompatActivity  {
 
     }
 
-    public static void SendData(long PacketID, ArrayStream Pack)
+    public static void SendData(int PacketID, ArrayStream Pack)
     {
         if(bStreamEnable)
         {
